Update exercise4.ipynb
Please refer to this link of Coursera discussion forum. [https://www.coursera.org/learn/machine-learning/discussions/weeks/5/threads/ag_zHUGDEeaXnBKVQldqyw]. In Q5 it is written that if we are doing element wise multiplication with g'(z), while calculating delta_2 , we must remove the first column of the Theta_2. 
The bug was that, it was asked to remove the first element of delta_2, which should be done only if we are doing element wise multiplication with a_2*(1- a_2), but here we are doing with g'(z).